Added an ABC linter
Summary:
https://trello.com/c/dku82NrD/4-detect-declarations-not-listed-alphabetically

This changeset adds our first linter, which checks that properties are listed
alphabetically (in accordance with our style guide).

For example, the following Less code violates our style guide:

```
a {
    width: 40px;
    color: red;
}
```

because "width" and "color" are not arranged alphabetically.

The linter currently ignores variables and mixins, and is only concerned with
properties. There is a TODO in place to address that.

Test Plan: npm test

Differential Revision: https://phabricator.khanacademy.org/D21237