Added a linter for overqualified selectors
Summary:
https://trello.com/c/zOipO85n/14-detect-element-id-overqualified-selectors

This changeset adds a linter that checks for overqualified selectors in the
form of `element#id` or `element.class`.

This check is done from the compiled CSS, meaning it can still catch errors
from code like:

```
ul {
    &.nav {
        // ...
    }
}
```

To do this, I first render the less code, then traverse the CSS AST with
PostCSS.

Also included in this changeset are changes to the unit tests, in which I
trim the Less code "fixtures" before testing, to have more intuitive line
numbers.

Test Plan: npm test

Differential Revision: https://phabricator.khanacademy.org/D21366