Prevent the overqualified linter from reporting things it's not supposed to
Summary:
https://trello.com/c/gjUn0tbo/29-prevent-percentages-from-being-classified-as-selectors

This changeset applies to fixes to the overqualified linter. First, the linter
no longer picks up `@keyframes` percentages as selectors and reports them as
being over qualified.

Additionally, the linter also will no longer report overqualified selectors
from imported files. They are tricky to report, and the linter should really
only concern itself with the file its run on.

Quite curiously, I also wrapped the callback invocation for the overqualified
linter with a `process.nextTick`. This forces the function to be asynchronous,
which is how it's used in the tests. Without it, I was noticing a race
condition in the asynchronous mocha tests.

Test Plan: npm test

Reviewers: emily

Differential Revision: https://phabricator.khanacademy.org/D21766