Fixed bug where nested @media queries caused strange output for the nesting linter
Summary:
https://trello.com/c/T1kcLD8y/36-fix-bug-with-nested-media-queries

This changeset adds a sanity check to the nesting linter to make sure we can
correctly point to a nesting violation.

It turns out that when we nest a @media query, the "index" property is
missing from the "element" field. Since @media queries do not contribute to
specificity, I am okay with simply ignoring them.

Test Plan: . npm test

Reviewers: emily

Differential Revision: https://phabricator.khanacademy.org/D21925