Display specific error messages for cases when @imported files do not parse
Summary:
https://trello.com/c/VyQZ2qUe/38-error-messages-for-imported-files

The last major pain point for integrating lesshint into khan-linter is that
if a file @imports another file which does not parse, the line numbers do not
correspond to the @import line but instead the part in the @imported file
which fails to parse.

This causes a mismatch in line numbers, and, in some cases, errors when we
attempt to display the violating line.

This changeset uses `(1:1)` as a location for parsing errors in @imported
files. There is a TODO to get the line/col for the actual `@import` statement.

Test plan:
n/a

Reviewers: emily